v1.1.0
* Add cluster functionality
* Extract core functions into `lib/app.js`
* Update devDependencies
