+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ip2country-bot",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Messenger Bot for IP 2 Country",
   "main": "index.js",
   "private": true,
@@ -14,8 +14,8 @@
     "messenger-bot": "^2.3.0"
   },
   "devDependencies": {
-    "config-leaf": "^0.2.0",
-    "eslint": "^3.1.1"
+    "config-leaf": "^0.3.0",
+    "eslint": "^3.3.0"
   },
   "scripts": {
     "test": "npm outdated && eslint .",
